Moved resource production tests to its package
